BACKPORT: drm/i915: only disable enabled planes on intel_fb_restore_mode
We should avoid touching registers that are on the power down well
when we don't need to, because if we touch these registers when the
power well is disabled we'll get tons of "unclaimed register"
messages. This commit fixes some of these messages.

Signed-off-by: Paulo Zanoni <paulo.r.zanoni@intel.com>
Signed-off-by: Daniel Vetter <daniel.vetter@ffwll.ch>

BUG=chromium:336809
TEST=Tested on snow & peppy

(cherry picked from commit 56e5a3f043f0440cc0cce288e71134d1cb1b44a7)
Signed-off-by: Sean Paul <seanpaul@chromium.org>