UPSTREAM: drm: add atomic fxns
The 'atomic' mechanism allows for multiple properties to be updated,
checked, and commited atomically.  This will be the basis of atomic-
modeset and nuclear-pageflip.

The basic flow is:

   state = dev->atomic_begin();
   for (... one or more ...)
      obj->set_property(obj, state, prop, value);
   if (dev->atomic_check(state))
      dev->atomic_commit(state);
   dev->atomic_end(state);

The split of check and commit steps is to allow for ioctls with a
test-only flag (which would skip the commit step).

Signed-off-by: Rob Clark <robdclark@gmail.com>

BUG=chromium:336809
TEST=Tested on snow & peppy

(cherry picked from commit 6dcab9fdbd2d66c13c5351ae3fc64899df02f275)
Signed-off-by: Sean Paul <seanpaul@chromium.org>

Conflicts:
	drivers/gpu/drm/Makefile
	drivers/gpu/drm/drm_crtc.c
	drivers/gpu/drm/exynos/exynos_drm_plane.c
	drivers/gpu/drm/gma500/psb_drv.c
	drivers/gpu/drm/msm/msm_drv.c
	drivers/gpu/drm/msm/msm_drv.h
	drivers/gpu/drm/vmwgfx/vmwgfx_drv.c

Change-Id: I4dbfad93a081d78ce67d4733ec3558833036c4b9