drm/exynos: Add a hack to appease userspace & atomic helpers
This patch adds a small hack to the plane initialization of the
exynos planes such that the "default window" has a possible_crtcs
of 0 and an overlay plane type. This ensures that userspace will
not select the plane for cursor use, but allows the planes to be
included in the drm layer plane loops which disable planes.

This will be removed once we no longer use the primary plane helpers
and properly implement it ourselves.

BUG=chromium:336809
TEST=Tested on snow & peppy

Change-Id: I0f22326935572617259e7f3b6a0e5c6bafd492b2